requirement.txt
pytest
selenium
openpyxl
allure-pytest
